Rename misc util to network util
