+++ elements/lisk-p2p/src/utils/network.ts
